+++ src/main/java/jmu/vo/Orders.java
@@ -9,7 +9,7 @@ public class Orders {
     private  String waybill_id;
     private  String  create_time;
     private   String  order_status;
-    private   float order_totalprice;
+    private   double order_totalprice;
     private   String receipt_status;
     private   String shipment_status;
 }
